+++ src/sagemaker/session.py
@@ -3663,6 +3663,7 @@ class Session(object):  # pylint: disable=too-many-public-methods
         volume_size=None,
         model_data_download_timeout=None,
         container_startup_health_check_timeout=None,
+        explainer_config_dict=None,
     ):
         """Create an Amazon SageMaker endpoint configuration.
 
@@ -3696,6 +3697,8 @@ class Session(object):  # pylint: disable=too-many-public-methods
                 inference container to pass health check by SageMaker Hosting. For more information
                 about health check see:
                 https://docs.aws.amazon.com/sagemaker/latest/dg/your-algorithms-inference-code.html#your-algorithms-inference-algo-ping-requests
+            explainer_config_dict (dict): Specifies configuration to enable explainers.
+                Default: None.
 
         Example:
             >>> tags = [{'Key': 'tagname', 'Value': 'tagvalue'}]
@@ -3751,6 +3754,9 @@ class Session(object):  # pylint: disable=too-many-public-methods
             )
             request["DataCaptureConfig"] = inferred_data_capture_config_dict
 
+        if explainer_config_dict is not None:
+            request["ExplainerConfig"] = explainer_config_dict
+
         self.sagemaker_client.create_endpoint_config(**request)
         return name
 
@@ -3762,6 +3768,7 @@ class Session(object):  # pylint: disable=too-many-public-methods
         new_kms_key=None,
         new_data_capture_config_dict=None,
         new_production_variants=None,
+        new_explainer_config_dict=None,
     ):
         """Create an Amazon SageMaker endpoint configuration from an existing one.
 
@@ -3789,6 +3796,9 @@ class Session(object):  # pylint: disable=too-many-public-methods
             new_production_variants (list[dict]): The configuration for which model(s) to host and
                 the resources to deploy for hosting the model(s). If not specified,
                 the ``ProductionVariants`` of the existing endpoint configuration is used.
+            new_explainer_config_dict (dict): Specifies configuration to enable explainers.
+                (default: None). If not specified, the explainer configuration of the existing
+                endpoint configuration is used.
 
         Returns:
             str: Name of the endpoint point configuration created.
@@ -3855,6 +3865,13 @@ class Session(object):  # pylint: disable=too-many-public-methods
                 sagemaker_session=self,
             )
             request["AsyncInferenceConfig"] = inferred_async_inference_config_dict
+
+        request_explainer_config_dict = (
+            new_explainer_config_dict or existing_endpoint_config_desc.get("ExplainerConfig", None)
+        )
+
+        if request_explainer_config_dict is not None:
+            request["ExplainerConfig"] = request_explainer_config_dict
 
         self.sagemaker_client.create_endpoint_config(**request)
 
@@ -4372,6 +4389,7 @@ class Session(object):  # pylint: disable=too-many-public-methods
         wait=True,
         data_capture_config_dict=None,
         async_inference_config_dict=None,
+        explainer_config_dict=None,
     ):
         """Create an SageMaker ``Endpoint`` from a list of production variants.
 
@@ -4388,6 +4406,9 @@ class Session(object):  # pylint: disable=too-many-public-methods
                 capture for use with Amazon SageMaker Model Monitoring. Default: None.
             async_inference_config_dict (dict) : specifies configuration related to async endpoint.
                 Use this configuration when trying to create async endpoint and make async inference
+                (default: None)
+            explainer_config_dict (dict) : Specifies configuration related to explainer.
+                Use this configuration when trying to use online explainability.
                 (default: None)
         Returns:
             str: The name of the created ``Endpoint``.
@@ -4422,6 +4443,8 @@ class Session(object):  # pylint: disable=too-many-public-methods
                 sagemaker_session=self,
             )
             config_options["AsyncInferenceConfig"] = inferred_async_inference_config_dict
+        if explainer_config_dict is not None:
+            config_options["ExplainerConfig"] = explainer_config_dict
 
         LOGGER.info("Creating endpoint-config with name %s", name)
         self.sagemaker_client.create_endpoint_config(**config_options)